Refactor: #5. FAQ ì°파일 충돌 해결
reques- t, response 디렉토리 변경 -ㅑindent 수정 -print문 제거
