+++ module-api/src/main/java/kernel/jdon/moduleapi/domain/faq/dto/response/CreateFaqResponse.java
@@ -1,19 +1,19 @@
-package kernel.jdon.moduleapi.domain.faq.dto;
+package kernel.jdon.moduleapi.domain.faq.dto.response;
 
 import kernel.jdon.moduleapi.domain.faq.entity.Faq;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @Builder
-@AllArgsConstructor
-public class FindFaqResponse {
+public class CreateFaqResponse {
+	private Long id;
 	private String title;
 	private String content;
 
-	public static FindFaqResponse of(Faq faq) {
-		return FindFaqResponse.builder()
+	public static CreateFaqResponse of(Faq faq) {
+		return CreateFaqResponse.builder()
+			.id(faq.getId())
 			.title(faq.getTitle())
 			.content(faq.getContent())
 			.build();
