+++ module-api/src/main/java/kernel/jdon/moduleapi/domain/faq/dto/response/FindFaqResponse.java
@@ -1,19 +1,19 @@
-package kernel.jdon.moduleapi.domain.faq.dto;
+package kernel.jdon.moduleapi.domain.faq.dto.response;
 
 import kernel.jdon.moduleapi.domain.faq.entity.Faq;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
 @Getter
 @Builder
-public class CreateFaqResponse {
-	private Long id;
+@AllArgsConstructor
+public class FindFaqResponse {
 	private String title;
 	private String content;
 
-	public static CreateFaqResponse of(Faq faq) {
-		return CreateFaqResponse.builder()
-			.id(faq.getId())
+	public static FindFaqResponse of(Faq faq) {
+		return FindFaqResponse.builder()
 			.title(faq.getTitle())
 			.content(faq.getContent())
 			.build();
